[Experiment]: Testing getLegAngle() method & Solve bugs in the stucking detection
The experiment is tested by manually control the rod, hence, a supporting strucutre
has been created to support the rod in the air to make move freely to test the angle.
For that, A bug has been known in the stucking detection:
When the controllers of the strings stuck because of the target length
is not feasible to be achieved. The bug was that if the controller stucked
the controller will not be able to extend the string again, because the wrong condition